Patched tracker status retrieval
The status of the tracker was not properly retrieved. It just tried to
connect to it, but sometimes the tracker returns "The tracker is down"
in its message to tell us that it doesn't work. A request similiar to
the normal site request to What.CD is now made to check on tracker
errors.